refactor: add ClaimsService
