+++ src/OidcClient.ts
@@ -14,6 +14,7 @@ import { SignoutResponse } from "./SignoutResponse";
 import { SigninState } from "./SigninState";
 import { State } from "./State";
 import { TokenClient } from "./TokenClient";
+import { ClaimsService } from "./ClaimsService";
 
 /**
  * @public
@@ -81,6 +82,7 @@ export class OidcClient {
     protected readonly _logger = new Logger("OidcClient");
 
     public readonly metadataService: MetadataService;
+    protected readonly _claimsService: ClaimsService;
     protected readonly _validator: ResponseValidator;
     protected readonly _tokenClient: TokenClient;
 
@@ -88,7 +90,8 @@ export class OidcClient {
         this.settings = new OidcClientSettingsStore(settings);
 
         this.metadataService = new MetadataService(this.settings);
-        this._validator = new ResponseValidator(this.settings, this.metadataService);
+        this._claimsService = new ClaimsService(this.settings);
+        this._validator = new ResponseValidator(this.settings, this.metadataService, this._claimsService);
         this._tokenClient = new TokenClient(this.settings, this.metadataService);
     }
